Use explicit type in some places and polish test names
